+++ pyron-core/src/main/scala/com/cloudentity/pyron/domain/rule/rule.scala
@@ -6,6 +6,11 @@ import com.cloudentity.pyron.domain.http.CallOpts
 case class RequestPluginsConf(pre: List[ApiGroupPluginConf], endpoint: List[ApiGroupPluginConf], post: List[ApiGroupPluginConf]) extends PluginsConf
 case class ResponsePluginsConf(pre: List[ApiGroupPluginConf], endpoint: List[ApiGroupPluginConf], post: List[ApiGroupPluginConf]) extends PluginsConf
 
+sealed trait BodyHandling
+  case object BufferBody extends BodyHandling
+  case object PipeBody extends BodyHandling
+  case object DropBody extends BodyHandling
+
 case class RuleConf(
   endpointName: Option[String],
   criteria: EndpointMatchCriteria,
@@ -16,6 +21,8 @@ case class RuleConf(
   copyQueryOnRewrite: Option[Boolean],
   preserveHostHeader: Option[Boolean],
   tags: List[String],
+  requestBody: Option[BodyHandling],
+  requestBodyMaxSize: Option[Int],
   call: Option[CallOpts],
   ext: ExtRuleConf
 )