+++ pyron-engine/src/main/scala/com/cloudentity/pyron/api/body/ContentLengthUtil.scala
@@ -1,0 +1,6 @@
+package com.cloudentity.pyron.api.body
+
+object ContentLengthUtil {
+  def isBodyLimitExceeded(contentLength: Int, maxBodySizeOpt: Option[Int]): Boolean =
+    maxBodySizeOpt.map(contentLength >> 10 > _).getOrElse(false) // contentLength / 1024 > maxBodySize
+}