+++ pyron-engine/src/main/scala/com/cloudentity/pyron/client/TargetClient.scala
@@ -2,8 +2,8 @@ package com.cloudentity.pyron.client
 
 import com.cloudentity.pyron.config.Conf
 import com.cloudentity.pyron.domain._
-import com.cloudentity.pyron.domain.flow.{DiscoverableService, DiscoverableServiceRule, FixedHttpClientConf, ServiceClientName, SmartHttpClientConf, StaticService, TargetHost}
-import com.cloudentity.pyron.domain.http.{CallOpts, TargetRequest}
+import com.cloudentity.pyron.domain.flow.{DiscoverableService, DiscoverableServiceRule, FixedHttpClientConf, Properties, RequestCtx, ServiceClientName, SmartHttpClientConf, StaticService, TargetHost}
+import com.cloudentity.pyron.domain.http.{CallOpts, Headers, TargetRequest}
 import com.cloudentity.pyron.domain.rule.RuleConf
 import com.cloudentity.tools.vertx.conf.ConfService
 import com.cloudentity.tools.vertx.http.builder.SmartHttpClientBuilderImpl.CallOk
@@ -20,6 +20,7 @@ import com.cloudentity.tools.vertx.sd.{Location, ServiceName}
 import io.vertx.core
 import io.vertx.core.Vertx
 import io.vertx.core.http._
+import io.vertx.core.streams.Pipe
 import org.slf4j.LoggerFactory
 import scalaz._
 
@@ -94,16 +95,16 @@ class TargetClient(tracing: TracingManager, fixedClient: HttpClient, clients: Ma
                   (implicit ec: VertxExecutionContext) extends FutureConversions {
   val log: LoggingWithTracing = LoggingWithTracing.getLogger(this.getClass)
 
-  def call(ctx: TracingContext, request: TargetRequest, callOpts: Option[CallOpts]): Future[Throwable \/ TargetResponse] = {
+  def call(ctx: TracingContext, request: TargetRequest, bodyPipeOpt: Option[Pipe[Buffer]], callOpts: Option[CallOpts]): Future[Throwable \/ TargetResponse] = {
     log.debug(ctx, s"Forwarding request: $request")
     val childCtx = buildChildContext(ctx, request)
     val requestWithTracingHeaders = injectTracingHeaders(childCtx, request)
 
     val result = request.service match {
       case StaticService(host, port, ssl) =>
-        callStaticService(ctx, Location(host.value, port, ssl, None), requestWithTracingHeaders, callOpts)
+        callStaticService(ctx, Location(host.value, port, ssl, None), requestWithTracingHeaders, bodyPipeOpt, callOpts)
       case DiscoverableService(serviceName) =>
-        callDiscoverableService(ctx, serviceName, requestWithTracingHeaders, callOpts)
+        callDiscoverableService(ctx, serviceName, requestWithTracingHeaders, bodyPipeOpt, callOpts)
     }
 
     result.onComplete {
@@ -138,25 +139,27 @@ class TargetClient(tracing: TracingManager, fixedClient: HttpClient, clients: Ma
     case DiscoverableService(serviceName) => serviceName.value
   }
 
-  private def callDiscoverableService(tracing: TracingContext, serviceName: ServiceClientName, request: TargetRequest, callOpts: Option[CallOpts]): Future[Throwable \/ TargetResponse] = {
+  private def callDiscoverableService(tracing: TracingContext, serviceName: ServiceClientName, request: TargetRequest, bodyPipeOpt: Option[Pipe[Buffer]], callOpts: Option[CallOpts]): Future[Throwable \/ TargetResponse] = {
     clients.get(serviceName) match {
-      case Some(cli) => makeSmartCall(tracing, request, callOpts, None, cli)
+      case Some(cli) => makeSmartCall(tracing, request, bodyPipeOpt, callOpts, None, cli)
       case None      => Future.failed(new Exception(s"Could not get SmartHttpClient for $serviceName"))
     }
   }
 
-  private def makeSmartCall(tracing: TracingContext, request: TargetRequest, callOpts: Option[CallOpts], sd: Option[Sd], cli: SmartHttpClient) = {
-    val reqBuilder =
-      copyHeadersWithoutContentLength(request, cli.request(request.method, request.uri.value))
-
-    val reqBuilderWithOpts = withCallOpts(reqBuilder, callOpts)
+  private def makeSmartCall(tracing: TracingContext, request: TargetRequest, bodyPipeOpt: Option[Pipe[Buffer]], callOpts: Option[CallOpts], sd: Option[Sd], cli: SmartHttpClient) = {
+    val reqBuilder = withCallOpts(cli.request(request.method, request.uri.value), callOpts)
 
     val responseFut: Future[SmartHttpResponse] =
       request.bodyOpt match {
         case Some(body) =>
-          reqBuilderWithOpts.endWithBody(tracing, body).toScala()
+          copyHeadersWithoutContentLength(request.headers, reqBuilder).endWithBody(tracing, body).toScala()
         case None =>
-          reqBuilderWithOpts.endWithBody(tracing).toScala()
+          bodyPipeOpt match {
+            case Some(pipe) =>
+              copyHeaders(request.headers, reqBuilder).endWithBody(tracing, pipe).toScala()
+            case None =>
+              copyHeaders(request.headers, reqBuilder).endWithBody(tracing).toScala()
+          }
       }
 
     responseFut
@@ -179,7 +182,7 @@ class TargetClient(tracing: TracingManager, fixedClient: HttpClient, clients: Ma
         reqBuilder
     }
 
-  private def callStaticService(tracing: TracingContext, location: Location, request: TargetRequest, callOpts: Option[CallOpts]): Future[Throwable \/ TargetResponse] = {
+  private def callStaticService(tracing: TracingContext, location: Location, request: TargetRequest, bodyPipeOpt: Option[Pipe[Buffer]], callOpts: Option[CallOpts]): Future[Throwable \/ TargetResponse] = {
     val sd = new Sd {
       val sn = ServiceName(s"http${if (location.ssl) "s" else ""}://${location.host}:${location.port}")
       override def discover(): Option[SdNode] = Some(SdNode(sn, new NoopCB(sn.value), location))
@@ -187,29 +190,20 @@ class TargetClient(tracing: TracingManager, fixedClient: HttpClient, clients: Ma
       override def close(): core.Future[Unit] = core.Future.succeededFuture(())
     }
 
-    makeSmartCall(tracing, request, callOpts, Some(sd), new SmartHttpClientImpl(sd, fixedClient, 0, None, _ => CallOk, _ => true))
-  }
-
-  def copyHeadersWithoutContentLength(from: TargetRequest, to: HttpClientRequest): Unit = {
-    for {
-      (name, values) <- from.headers.toMap
-      value          <- values
-    } to.headers().add(name, value)
-
-    dropContentLengthHeader(to)
+    makeSmartCall(tracing, request, bodyPipeOpt, callOpts, Some(sd), new SmartHttpClientImpl(sd, fixedClient, 0, None, _ => CallOk, _ => true))
   }
 
   /**
     * Plugins could have changed original body, without adjusting Content-Length.
     * We drop Content-Length and let Vertx http client set it.
     */
-  private def dropContentLengthHeader(request: HttpClientRequest) =
-    request.headers().remove("Content-Length")
+  def copyHeadersWithoutContentLength(from: Headers, to: RequestCtxBuilder): RequestCtxBuilder =
+    copyHeaders(from.remove("Content-Length"), to)
 
-  def copyHeadersWithoutContentLength(from: TargetRequest, to: RequestCtxBuilder): RequestCtxBuilder = {
+  def copyHeaders(from: Headers, to: RequestCtxBuilder): RequestCtxBuilder = {
     val headers: Iterable[(String, String)] =
       for {
-        (name, values) <- from.headers.remove("Content-Length").toMap
+        (name, values) <- from.toMap
         value          <- values.headOption
       } yield (name, value)